Move GlobalRandom to separate file and delegate it to Random.Shared on .NET 6+
